requirements.txt
xarray>=2023.1.0
netCDF4>=1.6.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
xgboost>=2.0.0
lightgbm>=4.0.0
torch>=2.0.0
flask>=3.0.0
joblib>=1.3.0
pyyaml>=6.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
colorama>=0.4.6
